refactor DB management, including moving firebase initialization from app to here
